Rearranged the files a bit. Still not sure of the right hierarchy. Beginning to think that kotlin scripts should live beside their corresponding scenes.
