+++ gdj/org/liamjd/kuiper/screens/KuiperGame.gdj
@@ -1,9 +1,9 @@
 // THIS FILE IS GENERATED! DO NOT EDIT OR DELETE IT. EDIT OR DELETE THE ASSOCIATED SOURCE CODE FILE INSTEAD
 // Note: You can however freely move this file inside your godot project if you want. Keep in mind however, that if you rename the originating source code file, this file will be deleted and regenerated as a new file instead of being updated! Other modifications to the source file however, will result in this file being updated.
 
-registeredName = GameSetup
-fqName = org.liamjd.kuiper.GameSetup
-relativeSourcePath = src/main/kotlin/org/liamjd/kuiper/GameSetup.kt
+registeredName = KuiperGame
+fqName = org.liamjd.kuiper.screens.KuiperGame
+relativeSourcePath = src/main/kotlin/org/liamjd/kuiper/screens/KuiperGame.kt
 baseType = Node
 supertypes = [
     godot.Node,
@@ -12,8 +12,7 @@ supertypes = [
 	kotlin.Any
 ]
 signals = [
-    country_select_signal,
-	start_game_signal
+    end_turn_signal
 ]
 properties = [
     
@@ -21,6 +20,5 @@ properties = [
 functions = [
     _ready,
 	_process,
-	_on_location_list_item_selected,
-	_on_start_game_button_pressed
+	_on_quit_game
 ]
