+++ src/main/scenes/game.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=2 format=3 uid="uid://dhc5at6r5jl4n"]
 
-[ext_resource type="Script" path="res://src/main/kotlin/org/liamjd/kuiper/KuiperGame.kt" id="1_0est4"]
+[ext_resource type="Script" path="res://src/main/kotlin/org/liamjd/kuiper/screens/KuiperGame.kt" id="1_0est4"]
 
 [node name="CanvasLayer" type="CanvasLayer"]
 script = ExtResource("1_0est4")
