+++ src/main/scenes/main_menu.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=2 format=3 uid="uid://dwd5bpnoghggx"]
 
-[ext_resource type="Script" path="res://src/main/kotlin/org/liamjd/kuiper/MainMenu.kt" id="1_nfpdk"]
+[ext_resource type="Script" path="res://src/main/kotlin/org/liamjd/kuiper/screens/MainMenu.kt" id="1_nfpdk"]
 
 [node name="MainMenu" type="CanvasLayer"]
 script = ExtResource("1_nfpdk")
@@ -12,7 +12,6 @@ anchor_bottom = 1.0
 grow_horizontal = 2
 grow_vertical = 2
 color = Color(0.125911, 0.125911, 0.125911, 1)
-script = ExtResource("1_nfpdk")
 
 [node name="VBoxContainer" type="VBoxContainer" parent="."]
 anchors_preset = 8
